Extract FileDsl
